refactor: change path to login
change path to login by removing it from /settings. Admins can only access the page using /admin in the URL. Update imports and paths in other files accordingly, and move some files for folder structure
